Player manager

+++ src/php/com/tcshl/player/Player.php
@@ -11,9 +11,15 @@
  	private $playerFName;
  	private $playerLName;
  	private $playerSkillLevel;
+ 	private $playerSkillLevelSelect;
  	private $playerRegistrationID;
  	private $playerSeasonID;
+
+        // Misc Attributes
+        private $playerFormErrors;
+        private $playerFormSuccess;
  	
+        // Constructor
  	function __construct($playerID) {
  		$this->playerID=(int)$playerID;
  		if($this->playerID != 0) {
@@ -47,7 +53,7 @@
  		$query .= ' registrationId = '.$this->get_playerRegistrationID().',';
  		$query .= ' seasonId = '.$this->get_playerSeasonID();
  		$query .= ' WHERE playerID='.$this->get_playerID();
- 		  			
+
 		$result = mysql_query($query)
   					or die("sp_clubs (Line " . __LINE__ . "): " . mysql_errno() . ": " . mysql_error());  			
  	}
@@ -68,12 +74,99 @@
  	} 
  	
  	function delete_player() {
-		$query = 'DELETE FROM '.PLAYER.' WHERE playerID='.$this->get_playerID();
+                if($this->canDelete()) {
+                        $query = 'DELETE FROM '.PLAYER.' WHERE playerID='.$this->get_playerID();
 
-		$result = mysql_query($query)
-  					or die("sp_clubs (Line " . __LINE__ . "): " . mysql_errno() . ": " . mysql_error());
+                        mysql_query($query)
+                                                or die("sp_clubs (Line " . __LINE__ . "): " . mysql_errno() . ": " . mysql_error());
+
+                        if(validResult()) {
+                                $this->teamFormSuccess[]='Player '.$this->get_playerFName().' '.$this->get_playerLName().' deleted successfully!';
+                        } else {
+                                $this->teamFormErrors[]='Player '.$this->get_playerFName().' '.$this->get_playerLName().' NOT deleted!  Try again or notify TCSHL.com administrator.';
+                        }
+                } else {
+                        $this->teamFormErrors[]='Player '.$this->get_playerFName().' '.$this->get_playerLName().' CANNOT be deleted as they belong to a team!';
+                }
+
  	} 	
  	
+        // Determines if a player can be deleted
+        public function canDelete() {
+                $query = 'SELECT * FROM '.ROSTERSOFTEAMSOFSEASONS.' WHERE playerID='.$this->get_playerID();
+
+                $result = mysql_query($query)
+                                        or die("sp_clubs (Line " . __LINE__ . "): " . mysql_errno() . ": " . mysql_error());
+
+                if($result && mysql_num_rows($result) > 0) {
+                        return false;
+                } else {
+                        return true;
+                }
+        }
+
+        // TeamFormReposts
+        public function formReposts($smarty) {
+                if ($_POST) {
+                        if ($_POST['playerFName']) {
+                                $smarty->assign('pfn', $_POST['playerFName']);
+                        }
+                        if ($_POST['playerLName']) {
+                                $smarty->assign('pln', $_POST['playerLName']);
+                        }
+                }
+        }
+
+        // TeamFormValidation
+        public function formValidation() {
+                if ($_POST['playerFName']) {
+                        if (strlen($_POST['playerFName']) < 2) {
+                                $this->playerFormErrors[] = "Player First Name must be at least 2 characters long.";
+                        }
+                } else {
+                        $this->playerFormErrors[] = "Player First name is a required field";
+                }
+                if ($_POST['playerLName']) {
+                        if (strlen($_POST['playerLName']) < 2) {
+                                $this->playerFormErrors[] = "Player Last Name must be at least 2 characters long.";
+                        }
+                } else {
+                        $this->playerFormErrors[] = "Player Last name is a required field";
+                }
+
+                return $this->get_playerFormErrors();
+        }
+
+        //formProcessInsert
+        function formProcessInsert() {
+                $this->formProcess();
+                $this->insert_player();
+                return $this->get_playerFormErrors();
+
+        }
+
+        //formProcessUpdate
+        function formProcessUpdate() {
+                $this->formProcess();
+                $this->update_player();
+                return $this->get_playerFormErrors();
+        }
+
+        //formProcessUpdate
+        function formProcessDelete() {
+                $this->delete_player();
+                return $this->get_playerFormErrors();
+        }
+
+        //formProcess
+        private function formProcess() {
+                $this->set_playerFName($_POST['playerFName']);
+                $this->set_playerLName($_POST['playerLName']);
+                $this->set_playerSkillLevel($_POST['skilllevel']);
+                $this->set_playerRegistrationID($_POST['playerRegistrationId']);
+                $this->set_playerSeasonID($_POST['season']);
+        }
+
  	public function get_playerID() {
  		return $this->playerID;
  	}
@@ -159,5 +252,25 @@
  		}
  		return $seasonNames;		
  	} 	
+
+        // Get player skill level select menu
+ 	public function get_playerSkillLevelSelect() {
+ 		return select_skill_level($this->playerSkillLevel);
+ 	} 
+
+        // Get player season select menu
+ 	public function get_playerSeasonSelect() {
+ 		return select_season($this->playerSeasonID);
+ 	} 
+
+        // Get $playerFormErrors
+        public function get_playerFormErrors() {
+                return (array) $this->playerFormErrors;
+        }
+
+        // Get $playerFormSuccess
+        public function get_playerFormSuccess() {
+                return (array) $this->playerFormSuccess;
+        }
  }
 ?>
